css+fixed bugs

+++ src/components/Main.js
@@ -1,19 +1,23 @@
-import {useState,useEffect} from 'react'
+import {useState,useEffect, useCallback} from 'react'
 import { Routes, Route } from 'react-router-dom';
 import Refrigerator from "../pages/Refrigerator"
 
-function Main(){
+function Main (props) {
     // import ingredient from backend
     const API_URL = "https://whatsfordinnerteam.herokuapp.com/refrigerator"
     const [refrigerator,setRefrigerator]=useState(null)
 
-    const getIngredient= async()=>{
+    const getIngredient= useCallback (async()=>{
         try{
-            const reponse = await fetch(API_URL,{
-                method:'GET'
-            })
-            const ingredients = await reponse.json();
-        setRefrigerator(ingredients)
+          const token = await props.user.getIdToken();
+            const response = await fetch(API_URL,{
+                method:'GET',
+                headers: {
+                  'Authorization': 'Bearer ' + token
+                }
+            });
+            const ingredients = await response.json();
+        setRefrigerator(ingredients);
         }catch(error){
             console.log(error);
         }
@@ -21,38 +25,59 @@ function Main(){
 
     const createIngredient = async (ingredient) => {
         try {
+          if (props.user) {
+            const token = await props.user.getIdToken();
             await fetch(API_URL, {
                   method: 'POST',
                   headers: {
                     'Content-Type': 'Application/json',
+                    'Authorization': 'Bearer ' + token
                   },
                   body: JSON.stringify(ingredient),
                 });
                 getIngredient();
+          }
             } catch (error) {
              console.log(error);         
             }
-          }
+    }
+
           const deleteIngredient = async (id) => {
-            await fetch(API_URL +"/"+id, {
-              method: 'DELETE',
-            });
-            getIngredient();
-          };
+            if (props.user) {
+              const token = await props.user.getIdToken();
+              await fetch(API_URL +"/"+id, {
+                method: 'DELETE',
+                headers: {
+                  'Authorization': 'Bearer ' + token
+                }
+              });
+            }
+              getIngredient();
+          }
+
           const editIngredient = async(ingredient,id)=>{
-            await fetch(API_URL+"/"+id,{
-            method:'PUT',
-            header:{
-                'Conmtent-Type':'Application/json'
-            },
-            body:JSON.stringify(ingredient),
-            })
+            if (props.user) {
+              const token = await props.user.getIdToken;
+              await fetch(API_URL+"/"+id,{
+              method:'PUT',
+              header:{
+                  'Conmtent-Type':'Application/json',
+                  'Authorization': 'Bearer ' + token
+              },
+              body:JSON.stringify(ingredient),
+              });
+            }
             getIngredient();
           }
 
 useEffect(()=>{
+  if (props.user) {
     getIngredient();
-}, []);
+  } else {
+    setRefrigerator(null);
+  }
+}, [props.user, getIngredient]);
+
 return(
     <main>
         <Routes>
