+++ src/components/RecipesMain.js
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 import { Routes, Route } from 'react-router-dom';
 
 import RecipeIndex from '../pages/RecipeIndex';
@@ -7,44 +7,55 @@ import RecipeUpdate from '../pages/RecipeUpdate';
 import RecipeNew from '../pages/RecipeNew';
 import { async } from '@firebase/util';
 
-function RecipesMain () {
-
+function RecipesMain (props) {
     const [savedRecipes, setSavedRecipes] = useState(null);
-
-    const recipesURL = 'http://localhost:2000/recipes/';
-
     const fetchSavedRecipes = useCallback(async () => {
         try {
-            const response = await fetch(props.recipesURL);
+            const token = await props.user.getIdToken();
+            const response = await fetch(props.recipesURL, {
+                method: 'GET',
+                headers: {
+                    'Authorization': 'Bearer ' + token
+                    }
+                });
             const data = await response.json();
             setSavedRecipes(data);
         } catch (error) {
             console.log(error);
         }
-    }
+    }, [props.user]);
 
     async function deleteSavedRecipe (id) {
-        await fetch(props.recipesURL + id, {
-            method: 'DELETE',
-        });
+        if (props.user) {
+            const token = await props.user.getIdToken();
+            await fetch(props.recipesURL + id, {
+                method: 'DELETE',
+                headers: {
+                    'Authorization': 'Bearer ' + token
+                }
+            });
+        }
         fetchSavedRecipes();
     }
 
     async function updateSavedRecipe (recipe, id) {
-        await fetch(props.recipesURL + id, {
-            method: 'PUT',
-            headers: {
-                'Content-Type': 'Application/json'
-            },
-            body: JSON.stringify(recipe)
-        });
+        if (props.user) {
+            const token = await props.user.getIdToken();
+            await fetch(props.recipesURL + id, {
+                method: 'PUT',
+                headers: {
+                    'Content-Type': 'Application/json',
+                    'Authorization': 'Bearer ' + token
+                },
+                body: JSON.stringify(recipe)
+            });
+        }
         fetchSavedRecipes();
     }
 
     function handleEatenTodayClick (id) {
 
         setSavedRecipes((savedRecipes) => {
-            // console.log(savedRecipes);
             const recipesCopy = [...savedRecipes];
             let foundIndex = recipesCopy.findIndex((r) => {
                 return r._id === id
